Fix mse_loss_indexes dtype when empty (torch.tensor([]) defaults to float)
Mixed VCoT+AO batches can produce empty mse_loss_indexes when a
GPU's pack has no output images. torch.tensor([]) creates float32,
causing IndexError. Force dtype=torch.long.

+++ data/dataset_base.py
@@ -232,7 +232,7 @@ class PackedDataset(torch.utils.data.IterableDataset):
         # if the model is required to perform visual generation
         if len(sequence_status['packed_timesteps']) > 0:
             data['packed_timesteps'] = torch.tensor(sequence_status['packed_timesteps'])
-            data['mse_loss_indexes'] = torch.tensor(sequence_status['mse_loss_indexes'])
+            data['mse_loss_indexes'] = torch.tensor(sequence_status['mse_loss_indexes'], dtype=torch.long)
 
         # if the model is required to perform text generation
         if len(sequence_status['packed_label_ids']) > 0:
